LPS-5984
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@40901 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/deploy/DeployUtil.java
@@ -147,18 +147,20 @@ public class DeployUtil {
 			tmpDir + "/liferay/com/liferay/portal/deploy/dependencies/" +
 				resource);
 
-		File parentFile = file.getParentFile();
+		if (!file.exists()) {
+			File parentFile = file.getParentFile();
 
-		if (parentFile != null) {
-			parentFile.mkdirs();
+			if (parentFile != null) {
+				parentFile.mkdirs();
+			}
+
+			byte[] bytes = FileUtil.getBytes(
+				getClass().getResourceAsStream("dependencies/" + resource));
+
+			OutputStream os = new FileOutputStream(file);
+
+			os.write(bytes);
 		}
-
-		byte[] bytes = FileUtil.getBytes(
-			getClass().getResourceAsStream("dependencies/" + resource));
-
-		OutputStream os = new FileOutputStream(file);
-
-		os.write(bytes);
 
 		return FileUtil.getAbsolutePath(file);
 	}
